Ansätze

QuantVarEntHam.H_A_Var — Type

H_A_Var

Struct to save the Yao AbstractBlocks and its matrix representation throughout the optimization.

The optimizing is only done with the matrices in matrices. blocks is only saved for convenient utiliy functions such as printing the Entanglement Hamiltonian.

Fields

blocks::Vector{AbstractBlock}: Containing each Hamiltonian Block as an Yao AbstractBlock.

matrices::Vector{Matrix{ComplexF64}}: Containing each Hamiltonian Block as a matrix.

QuantVarEntHam.H_A_BW — Method

H_A_BW(set::Settings)

Return an instance of type H_A_Var containing the Yao Blocks and its matrix representation.

The variational Ansatz follows the Bisognano-Wichmann-theorem. This function dispatches on the concrete subtypes of the abstract type Settings to get the correct variational Ansatz for the corresponding model.

Example

H_A_BW(set::Settings_TFIM) returns the variational Ansatz for the TFIM following the Bisognano-Wichmann-theorem.

QuantVarEntHam.H_A_not_BW — Method

H_A_not_BW(set::Settings)

Return an instance of type H_A_Var containing the Yao Blocks and its matrix representation.

The variational Ansatz does not follow the Bisognano-Wichmann-theorem. This function dispatches on the concrete subtypes of the abstract type Settings to get the correct variational Ansatz for the corresponding model.

Example

H_A_not_BW(set::Settings_TFIM) returns the variational Ansatz for the TFIM following not following the Bisognano-Wichmann-theorem.