Settings

QuantVarEntHam.Settings — Type

Settings{T<:AbstractBlock,S<:AbstractMatrix}

Abstract type to dispatch on the concrete types for the correct variational Ansätze.

The parametric type T<:AbstractBlock is introduced for determining the correct concrete type of the Yao Blocks, while S<:AbstractMatrix is needed to determine the concrete type of the matrix representation of the Yao Blocks to prevent working with full complex dense matrices.

QuantVarEntHam.Settings_XXZ — Type

Settings_XXZ{T<:AbstractBlock, S<:AbstractMatrix} <:Settings{T,S}

Contains the settings for the XXZ model

The parametric type T<:AbstractBlock is introduced for determining the correct concrete type of the Yao Blocks, while S<:AbstractMatrix is needed to determine the concrete type of the matrix representation of the Yao Blocks to prevent working with full complex dense matrices.

Tip

Use the constructor XXZ to instantiate this struct since the types for observables and its matrices mtrxObs are automatically inferred then and the default values in XXZ are highly recommended.

Fields

N::Int: number of sites in composite system.

N_A::Int: number of sites in subsystem A.

Δ::Float64: Anisotropy.

T_max::Float64: maximum time for evolving the observables i.e. maximum integration time.

r_max::Int: range of interaction (1 for nearest neighbour, 2 for next nearest neighbour, etc..) rmax = NA-1 corresponds to maximum order.

periodic::Bool: boundary conditions for the system Hamitlonian, false for open and true for periodic boundary conditions, obsolete if an own reduced density matrix ρ_A is provided.

ρ_A::DensityMatrix{2, ComplexF64, Matrix{ComplexF64}}: reduced density matrix of ground state of the composite system on subsystem A.

observables::Vector{T}: monitored observables in the cost function.

meas0::Vector{Float64} = [expect(observables[i], ρ_A) for i in eachindex(observables)]: expectation values of observables at time $t=0$.

mtrxObs::Vector{S}: matrix representations for the observables

atol::Float64=0.0: absolute tolerance for the integrator.

rtol::Float64=atol>0 ? 0.: sqrt(eps(Float64)): relative tolerance for the integrator.

dt::Float64=0.01: time step for evaluating the cost function via midpoint rule, obsolete if other integration techniques are used.

QuantVarEntHam.XXZ — Method

XXZ(N::Int, N_A::Int, Δ::Real, T_max::Real; r_max::Int=1, periodic::Bool = false, atol::Real=0.0, rtol::Real=atol>0 ? 0. : sqrt(eps(Float64)),
signHam::Integer=+1, ρ_A::DensityMatrix{2}=get_rhoA(H_XXZ(N, Δ, periodic=periodic, signHam=signHam), N-N_A+1:N, N),
observables::Vector{<:AbstractBlock}=[repeat(N_A, Z, (i,i+1)) for i in 1:N_A-1], dt::Float64 = 0.01)

Convenient constructor for Settings_XXZ containing settings for the XXZ Model

Required Arguments

N::Int: number of sites in the composite system.

Δ::Real: Anisotropy

N_A::Int: number of sites in subsystem A.

T_max::Real: maximum time for evolving the observables i.e. maximum integration time.

Keyword arguments

r_max::Int=1: range of interaction (1 for nearest neighbour, 2 for next nearest neighbour, etc..) rmax = NA-1 corresponds to maximum order.

periodic::Bool=false: boundary conditions for the system Hamitlonian, false for open and true for periodic boundary conditions, obsolete if an own reduced density matrix ρ_A is provided.

atol::Real=0.0: absolute tolerance for the integrator.

rtol::Real=atol>0 ? 0.: sqrt(eps(Float64)): relative tolerance for the integrator.

signHam::Integer=1: global sign of Hamiltonian, obsolete if an own reduced density matrix ρ_A is provided.

ρ_A::DensityMatrix{2}=get_rhoA(H_XXZ(N, Δ, periodic=periodic, signHam=signHam), N-N_A+1:N, N): reduced density matrix of ground state of the composite system on subsystem A, by default the subsystem is on the right border.

observables::Vector{<:AbstractBlock}=[repeat(N_A, Z, (i,i+1)) for i in 1:N_A-1]: monitored observables in the cost function.

dt::Real=0.01: time step for evaluating the cost function via midpoint rule, obsolete if other integration techniques are used.

Recommendations

use only Z-gates (or composition of these) as observables since these are diagonal thus save computation time the most.

use only one type of observables (e.g. Xi Xi+1) since these are then stored as sparse or diagonal matrices, otherwise dense which leads to higher computation time.

be carefull when changing the tolerances for integration (not recommended), a relative tolerance higher than ≈ 1e-7 is not recommended since this can lead to wrong results.

QuantVarEntHam.Settings_TFIM — Type

Settings_TFIM{T<:AbstractBlock, S<:AbstractMatrix} <:Settings{T,S}

Contains the settings for the TFIM

The parametric type T<:AbstractBlock is introduced for determining the correct concrete type of the Yao Blocks, while S<:AbstractMatrix is needed to determine the concrete type of the matrix representation of the Yao Blocks to prevent working with full complex dense matrices.

Tip

Use the constructor TFIM to instantiate this struct since the types for observables and its matrices mtrxObs are automatically inferred then and the default values in TFIM are highly recommended.

Fields

N::Int: number of sites in composite system.

N_A::Int: number of sites in subsystem A.

Γ::Float64: Anisotropy.

T_max::Float64: maximum time for evolving the observables i.e. maximum integration time.

r_max::Int: range of interaction (1 for nearest neighbour, 2 for next nearest neighbour, etc..) rmax = NA-1 corresponds to maximum order.

periodic::Bool: boundary conditions for the system Hamitlonian, false for open and true for periodic boundary conditions, obsolete if an own reduced density matrix ρ_A is provided.

ρ_A::DensityMatrix{2, ComplexF64, Matrix{ComplexF64}}: reduced density matrix of ground state of the composite system on subsystem A.

observables::Vector{T}: monitored observables in the cost function.

meas0::Vector{Float64} = [expect(observables[i], ρ_A) for i in eachindex(observables)]: expectation values of observables at time $t=0$.

mtrxObs::Vector{S}: matrix representations for the observables

atol::Float64=0.0: absolute tolerance for the integrator.

rtol::Float64=atol>0 ? 0.: sqrt(eps(Float64)): relative tolerance for the integrator.

dt::Float64=0.01: time step for evaluating the cost function via midpoint rule, obsolete if other integration techniques are used.

QuantVarEntHam.TFIM — Method

TFIM(N::Int, N_A::Int, Γ::Real, T_max::Real; r_max::Int=1, periodic::Bool=false, atol::Real=0.0, rtol::Real=atol>0 ? 0. : sqrt(eps(Float64)),
signHam::Integer=-1, ρ_A::DensityMatrix{2}=get_rhoA(H_TFIM(N, Γ, periodic = periodic, signHam=signHam), N-N_A+1:N, N),
observables::Vector{<:AbstractBlock}=[repeat(N_A, Z, (i,i+1)) for i in 1:N_A-1], dt::Real=0.01)

Convenient constructor for Settings_TFIM containing settings for the Transversal Field Ising Model

Required Arguments

N::Int: number of sites in the composite system.

Γ::Real: transversal field strength

N_A::Int: number of sites in subsystem A.

T_max::Real: maximum time for evolving the observables i.e. maximum integration time.

Keyword arguments

r_max::Int=1: range of interaction (1 for nearest neighbour, 2 for next nearest neighbour, etc..) rmax = NA-1 corresponds to maximum order.

periodic::Bool=false: boundary conditions for the system Hamitlonian, false for open and true for periodic boundary conditions.

atol::Real=0.0: absolute tolerance for the integrator.

rtol::Real=atol>0 ? 0.: sqrt(eps(Float64)): relative tolerance for the integrator.

signHam::Integer=-1: global sign of Hamiltonian, obsolete if an own reduced density matrix ρ_A is provided.

ρ_A::DensityMatrix{2}=get_rhoA(H_TFIM(N, Γ, periodic = periodic, signHam=signHam), N-N_A+1:N, N): reduced density matrix of ground state of the composite system on subsystem A, by default the subsystem is on the right border.

observables::Vector{<:AbstractBlock}=[repeat(N_A, Z, (i,i+1)) for i in 1:N_A-1]: monitored observables in the cost function.

dt::Real=0.01: time step for evaluating the cost function via midpoint rule, obsolete if other integration techniques are used.

Recommendations

use only Z-gates (or composition of these) as observables since these are diagonal thus save computation time the most.

use only one type of observables (e.g. Xi Xi+1) since these are then stored as sparse matrices, otherwise dense which leads to higher computation time.

be carefull when changing the tolerances for integration (not recommended), a relative tolerance higher than ≈ 1e-7 is not recommended since this can lead to wrong results.